Extend initialisation script to detect conda tab-completion and remove it if installed.
Conda tab expansion registers a PowerShell function to provide expansion of commands. However, the implementation is old and lacks extensive support for completions. It is automatically installed and there are no options to prevent installation. The script detects if conda is registered as the main tab-expansion function and deletes it if it is. It then re-instates any tab-expansion function that may have existed prior to conda installing its own tab-expansion function. NOTE: This function will not remove the conda tab-expansion if it is renamed by the subsequent installation of another tab-expansion script.