Implementation of basic tab-completion capability.
This commit establishes the basis framework for tab-completions. It only returns arguments and does not consider aliases. It doesn't consider whether the completion has already been enetered (and is, therefore, not appropriate as a suggestion). It does not provide any guidance on parameter values and suggest alternatives (e.g. conda environments). We are also missing conda commands not included within the documentation (conda activate/deactivate).