Initial code to execute commands within PowerShell.
Several commands have argument values which can be obtained by executing commands locally on the system. For example, the list of conda environments is likely to be small and could be suggested to the user during tab-completion. To gather the information commands need to be run within a PowerShell environment. This code creates a PowerShell instance and executes a command. It needs to be developed so that the code is robust and avoids deadlocks by timeing out gracefully if required (e.g. no command should take more than 10 mSec).