Move responsibility for creating parameter value suggestions record to helper.
Perviously, the helper only returned a list of strings. However, by moving the suggestion record creation to the helper we can specify the completion text, listing and tooltup allowing much greater flexibility in presenting information.

The helper is also now reponsible for filtering suggestions based upon the word to comlete, which is passed as a parameter to the helper method.