Code quality - Implement bool property to identify check type of SyntaxItem.
The code previously tested the type of the syntax item by comparing the text value of the Type property against a text string. This requires knowledge of the text values used to describe syntax item type and creates a dependency on the internal implementation of syntax items. Better to hide the implementation and expose a bool property IsXXXXX to indicate what type the syntax item is.