Bug fix: Implement constantExpressionAst now that default tokens are not added.
Adding tokens when the Visit default is called meant that tokens without specific handlers could be added. Once the action of the VisitDefault was changed to not add a token, to exclude tokens that were added by the PSReadLine resolver, the ability to record constant numbers was lost. Handling constantExpressionAst was added to identify constants (e.g. 12) on the command line.