Bug Fix - Add authorisatoin manager to initial session state in tests.
Testing on another machine the cmdlet tests do not work. This was caused when the authorisation manager refused to allow the PSReadLine module to be loaded. Adding a permissive authorisation manager (default with Microsoft.Powershel to the initial session state resolved this issue..

NOTE: If this test is performed in a continuous integration environment then the Authorisation Manager may inherit the execution policy from the PowerShell host. Therefore is Set-Execution policy is not permissive this test may fail in the CI environment, but work on machines where the execution policy is more permissive. The Microsoft document is not clear on inheritance when setting up the initial session state.