Bug fix: Predictor was not returning commands as suggestion
Set word to complete to blank if the last character is a space on the command line. Previously, the word to complete was the last token, which caused suggestions to be filtered by this word, with the result that no suggestions were made. A blank token was added to trigger suggestions, however, this caused commands to be filtered, but parameters were not -- need to check why parameters were not filtered as well.