Update conda tests to add conda environments to the shell prior to test execution.
Appvayor fails because the conda environment variable are not set when the conda validation tests run. This exposed a weakness in the testing strategy whereby the test had worked around the conda helper ENVIRONMENT not picking up that conda environment variables were not installed and providing a result regardless. In reality, the helper needs to identify that conda is not in the environment and return no suggestions. The test then needs to change to add the Conda Environment variables to the shell before running the test.

A further problem needs to be resolved: How do we find conda in the test system reliably so that we can call it during testing?