New feature: Cache tokeniser once parsed.
Tab-expansion reparses the command line. Stashing the tokeniser once it is parsed by the resolver allows tab-expansion to re-use it rather than parse the command line again. The tokeniser is stashed against the guid for the application. During testing the resolver runs in parallel with other tests and can stash a tokeniser which interferes with other tests. To mitigate this the test resolver is given a different guid and the tokeniser is stored against this temporary guid rather than the application guid.

During testing this feature it was noted that two tests were not 'using' powershell mock. This was amended to improve code consistency.