Testing: Add additional tests for the state machine.
Code Quality: Rename Token.Complete to Token.IsComplete for consistency.
Code Quality: Comment updates for clarity.

NOTE: Need to address the issue associated with running tests in a multi-threaded environment. This cases difficulty with the caching of values - the caches are static and become shared bewtween multiple threads which creates problems of coherence and consistency when two threads are updating the cache/tokeniser at the same time..